Move Chinese Wiktionary extract gloss code to a new "gloss.py" file
